Move soft error definitions to a separate header file
